ingest_app_store_data_to_bigquery/requirements.txt
#google-api-core==1.29.0
#google-api-core[grpc,grpcgcp]==1.14.0
#google-api-python-client==1.9.3
#google-auth==1.30.0
#google-auth-httplib2==0.0.4
#google-auth-oauthlib==0.4.1
#google-cloud==0.34.0
google-cloud-bigquery==1.25.0
#google-cloud-bigquery-storage==2.0.1
#google-cloud-bigtable==1.2.1
#google-cloud-core ==1.4.3
#google-cloud-datastore==1.12.0
#google-cloud-language==2.0.0
#google-cloud-logging==1.15.0
#google-cloud-pubsub==2.1.0
#google-cloud-resource-manager==0.30.2
#google-cloud-scheduler==2.2.0
#google-cloud-secret-manager==2.0.0
#google-cloud-spanner==1.19.1
google-cloud-storage==1.29.0
google-cloud==0.34.0
#grpc-google-iam-v1==0.12.3
#grpcio==1.29.0
#google-resumable-media

pandas==1.4.3
appstoreconnect==0.10.0